requirements.txt
Django==2.2.5
django-ckeditor==5.7.1
django-filebrowser==3.9.2
django-grappelli==2.10.4
django-js-asset==1.2.2
gunicorn==19.9.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
